+++ packages/wonder-blocks-form/src/components/__tests__/labeled-text-field.test.tsx
@@ -3,7 +3,6 @@ import {render, screen, fireEvent} from "@testing-library/react";
 import {userEvent} from "@testing-library/user-event";
 
 import {StyleSheet} from "aphrodite";
-import {color} from "@khanacademy/wonder-blocks-tokens";
 import LabeledTextField from "../labeled-text-field";
 
 describe("LabeledTextField", () => {
@@ -174,7 +173,7 @@ describe("LabeledTextField", () => {
 
         // Assert
         const input = await screen.findByRole("textbox");
-        expect(input).toBeDisabled();
+        expect(input).toHaveAttribute("aria-disabled", "true");
     });
 
     it("ariaDescribedby prop sets aria-describedby", async () => {
@@ -382,8 +381,9 @@ describe("LabeledTextField", () => {
         expect(input).toBeInTheDocument();
     });
 
-    it("light prop is passed to textfield", async () => {
+    it("name prop is passed to input", async () => {
         // Arrange
+        const name = "test-name";
 
         // Act
         render(
@@ -391,17 +391,13 @@ describe("LabeledTextField", () => {
                 label="Label"
                 value=""
                 onChange={() => {}}
-                light={true}
+                name={name}
             />,
         );
 
-        const textField = await screen.findByRole("textbox");
-        textField.focus();
-
         // Assert
-        expect(textField).toHaveStyle({
-            boxShadow: `0px 0px 0px 1px ${color.blue}, 0px 0px 0px 2px ${color.white}`,
-        });
+        const input = await screen.findByRole("textbox");
+        expect(input).toHaveAttribute("name", name);
     });
 
     it("style prop is passed to fieldheading", async () => {