+++ packages/wonder-blocks-testing/src/gql/mock-gql-fetch.ts
@@ -1,4 +1,7 @@
-import {mockRequester} from "@khanacademy/wonder-blocks-testing-core";
+import {
+    GraphQLJson,
+    mockRequester,
+} from "@khanacademy/wonder-blocks-testing-core";
 import {gqlRequestMatchesMock} from "./gql-request-matches-mock";
 import type {GqlFetchMockFn, GqlMockOperation} from "./types";
 
@@ -6,7 +9,7 @@ import type {GqlFetchMockFn, GqlMockOperation} from "./types";
  * A mock for the fetch function passed to GqlRouter.
  */
 export const mockGqlFetch = (): GqlFetchMockFn =>
-    mockRequester<GqlMockOperation<any, any, any>>(
+    mockRequester<GqlMockOperation<any, any, any>, GraphQLJson<any>>(
         gqlRequestMatchesMock,
         // Note that the identation at the start of each line is important.
         // TODO(somewhatabstract): Make a stringify that indents each line of