+++ packages/wonder-blocks-dropdown/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @khanacademy/wonder-blocks-dropdown
 
+## 5.6.0
+
+### Minor Changes
+
+-   0b3a28a7: - Combobox: Add error prop to support aria-invalid and styling changes.
+    -   TextField: Modify aria-invalid order to be overriden by the caller.
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-search-field@2.3.3
+
+## 5.5.6
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-search-field@2.3.2
+
+## 5.5.5
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-modal@5.1.14
+-   @khanacademy/wonder-blocks-search-field@2.3.1
+
+## 5.5.4
+
+### Patch Changes
+
+-   Updated dependencies [659a031d]
+    -   @khanacademy/wonder-blocks-search-field@2.3.0
+    -   @khanacademy/wonder-blocks-modal@5.1.13
+
+## 5.5.3
+
+### Patch Changes
+
+-   @khanacademy/wonder-blocks-search-field@2.2.27
+
 ## 5.5.2
 
 ### Patch Changes